Cast gsize to int for printf. Patch by Saleem Abdulrasool
2007-04-24  William Jon McCann  <mccann@jhu.edu>

	* common/test-config.c: (load_servers_group):
	Cast gsize to int for printf.
	Patch by Saleem Abdulrasool  <compnerd@gentoo.org>.
	Fixes bug #432830.


svn path=/trunk/; revision=4872

+++ common/test-config.c
@@ -163,7 +163,7 @@ load_servers_group (GdmConfig *config)
         int        i;
 
         keys = gdm_config_get_keys_for_group (config, GDM_CONFIG_GROUP_SERVERS, &len, NULL);
-        g_message ("Got %d keys for group %s", len, GDM_CONFIG_GROUP_SERVERS);
+        g_message ("Got %d keys for group %s", (int)len, GDM_CONFIG_GROUP_SERVERS);
 
         /* now construct entries for these groups */
         for (i = 0; i < len; i++) {
